+++ tests/testthat/test-data_dir.R
@@ -22,17 +22,3 @@ test_that("Get STATS19_DOWNLOAD_DIRECTORY ", {
   if(c == "") c = tempdir()
   expect_true(identical(get_data_directory(), c))
 })
-
-test_that("Reset STATS19_DOWNLOAD_DIRECTORY ", {
-  test_path = file.path(tempdir(), "stats19")
-  dir.create(test_path)
-  # hold value
-  c = get_data_directory()
-  Sys.setenv(STATS19_DOWNLOAD_DIRECTORY = "")
-  set_data_directory(test_path)
-  expect_true(identical(get_data_directory(), test_path))
-  # reset test user's prefs
-  if(!identical(c, tempdir())) {
-    set_data_directory(c)
-  }
-})