+++ NEWS.md
@@ -1,3 +1,7 @@
+# stats19 3.0.0 2023-10
+
+* Major update so the package works with the new csv files
+
 # stats19 2.0.1 2022-11
 
 * Changes spatstat.core related code (#217)